Move AgentOperationsManager into galaxy-app package
galaxy.agents.history imports AgentOperationsManager at module level,
but it lived in galaxy.webapps.galaxy.services which is in the
galaxy-webapps package. The galaxy-app package doesn't depend on
galaxy-webapps, so the package tests couldn't even collect -- 98 test
collection errors across all tests/app/ tests.

The manager's own top-level imports are all galaxy.managers/schema/
structured_app (already in galaxy-app), and its service-layer imports
are already lazy via properties, so moving it to galaxy.agents.operations
is a clean fit.
