+++ lib/galaxy/agents/operations.py
