+++ lib/galaxy/webapps/galaxy/api/mcp.py
@@ -15,8 +15,8 @@ from fastmcp import (
     settings as fastmcp_settings,
 )
 
+from galaxy.agents.operations import AgentOperationsManager
 from galaxy.managers.users import UserManager
-from galaxy.webapps.galaxy.services.agent_operations import AgentOperationsManager
 from galaxy.work.context import WorkRequestContext
 
 logger = logging.getLogger(__name__)
